requirements.txt
# Dragonfruit Proposal & MVP
openpyxl>=3.1.0
python-docx>=1.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.18.0

# Agentic MVP — OpenAI Agents SDK + connectors
openai-agents>=0.10.0
openai>=1.0.0
httpx>=0.27.0
python-dotenv>=1.0.0

# MCP: mock data server (expose mock_data/*.json to agents). Requires Python 3.10+.
# mcp[cli]>=1.0.0
